fix task 13

+++ python-object_relational_mapping/13-model_state_delete_a.py
@@ -20,10 +20,9 @@ if __name__ == "__main__":
 
     session = Session()
 
-    states = session.query(State).filter(State.name.ilike('%a%')).all()
-
-    for state in states:
-        session.delete(state)
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
 
     session.commit()
